requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2
gspread==6.1.2
google-auth==2.30.0
gunicorn==22.0.0
